+++ docs/changelog.rst
@@ -1,9 +1,12 @@
 Changelog
 =========
+
 0.2 - 2014-XX-XX
 ~~~~~~~~~~~~~~~~
 
-* In development.
+**In development**
+
+* Added initial CommonCrypto bindings.
 
 0.1 - 2014-01-08
 ~~~~~~~~~~~~~~~~